public double personalPerf = 0;// 个人表现

+++ src/netdata/UserReport.java
@@ -39,5 +39,5 @@ public class UserReport implements Serializable {
 	public int hidePackSucNum = 0;//埋包成功次数
 	public int openPackSucNum = 0;//拆包成功次数
 	public int roundNum = 0;//爆破模式的回合数
-	public int personalPerf = 0;// 个人表现
+	public double personalPerf = 0;// 个人表现
 }
